% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkingFunctions.R
\name{ppf.preferences}
\alias{ppf.preferences}
\title{Probabilistic Parking Functions}
\usage{
ppf.preferences(N, m, n, j, p)
}
\arguments{
\item{N}{number of repetitions}

\item{m}{Number of cars.}

\item{n}{Number of spots.}

\item{j}{j-th car.}

\item{p}{Probability of driving to the right.}
}
\value{
A list of preferences or NULL if in N repetitions there were no successful in parking
}
\description{
Get the preferences of the j-th car of a parking function
}
