% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkingFunctions.R
\name{ppf}
\alias{ppf}
\title{Probabilistic Parking Functions}
\usage{
ppf(m, n, p)
}
\arguments{
\item{m}{Number of cars.}

\item{n}{Number of spots.}

\item{p}{Probability of driving to the right.}
}
\value{
A list pref = preference vectors spots either the parked scheme or FALSE is some car didn't park
}
\description{
Run a uniformly distributed (m,n)-PPF
}
