% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkingFunctions.R
\name{ppf.park}
\alias{ppf.park}
\title{Probabilistic Parking Functions}
\usage{
ppf.park(prefs, n, p)
}
\arguments{
\item{n}{Number of spots.}

\item{p}{Probability of driving to the right.}

\item{m}{Number of cars.}
}
\value{
A vector with the park configuration. If the result has NA, then some car didn't park.
}
\description{
Try to park the cars given the preference vector
}
