% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkingFunctions.R
\name{ppf.parked}
\alias{ppf.parked}
\title{Probabilistic Parking Functions}
\usage{
ppf.parked(prefs, n, p)
}
\arguments{
\item{prefs}{Vector of cars' preferences}

\item{n}{Number of spots.}

\item{p}{Probability of driving to the right.}
}
\value{
TRUE if all cars have parked FALSE otherwise
}
\description{
Try to park all cars given their preferences, number of spots and direction probability
}
