% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parkingFunctions.R
\name{ppf.preferences}
\alias{ppf.preferences}
\title{Retrieve Preferences of the j-th Car in a Probabilistic Parking Function}
\usage{
ppf.preferences(N, m, n, j, p)
}
\arguments{
\item{N}{The number of repetitions for the simulation.}

\item{m}{The total number of cars.}

\item{n}{The total number of parking spots.}

\item{j}{The index of the car whose preference you want to retrieve (1 ≤ j ≤ m).}

\item{p}{The probability of a car deciding to move to the right if its preferred spot is occupied (0 ≤ p ≤ 1).}
}
\value{
A list of parking preferences for the j-th car across successful parking attempts, or `NULL` if no successful parking occurred in `N` repetitions.
}
\description{
This function simulates the parking process multiple times to collect the preferred parking spots of the j-th car, provided that the car successfully parks.
It attempts to park the cars in `N` independent repetitions and records the preference of the j-th car each time it successfully parks.
}
